Get rid of tests package and just use @ethdebug/format
