+++ packages/format/src/schemas/validity.test.ts
@@ -1,14 +1,28 @@
 import { describe, it } from "vitest";
 import {
+  type OutputUnit,
   validate,
   InvalidSchemaError,
 } from "@hyperjump/json-schema/draft-2020-12";
 
-import schemas from "../src/schemas.js";
-import printErrors from "../src/printErrors.js";
+import { schemas } from ".";
 
 // loads schemas into global hyperjump json schema validator
-import "../src/loadSchemas.js";
+import "../../test/hyperjump";
+
+const printErrors = (output: OutputUnit): string => output.errors!
+  .map((error) => {
+    if (!error.valid && !error.keyword.endsWith("#validate")) {
+      return `${
+        error.instanceLocation
+      } fails schema constraint ${
+        error.absoluteKeywordLocation
+      }`;
+    }
+  })
+  .filter((message): message is string => !!message)
+  .map(message => `  - ${message}`)
+  .join("\n");
 
 describe("Valid schemas", () => {
   for (const [id, schema] of Object.entries(schemas)) {
