+++ packages/format/vitest.d.ts
@@ -1,6 +1,9 @@
 import "vitest";
+import type { DescribeSchemaOptions } from "./src/describe";
 
 interface CustomMatchers<R = unknown> {
+  toValidate(schemaOptions: DescribeSchemaOptions): R;
+
   toSatisfy(received: any): R;
   toSatisfyAll(values: readonly any[]): R;
 }
