--- packages/tests/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@ethdebug/format-tests",
-  "version": "0.1.0-0",
-  "description": "Test harness for validating ethdebug/format schemas",
-  "main": "index.js",
-  "type": "module",
-  "license": "MIT",
-  "private": true,
-  "scripts": {
-    "test": "vitest"
-  },
-  "devDependencies": {
-    "@ethdebug/format": "^0.1.0-0",
-    "@hyperjump/browser": "^1.2.0",
-    "@hyperjump/json-schema": "^1.11.0",
-    "cli-highlight": "^2.1.11",
-    "indent-string": "^5.0.0",
-    "typescript": "^5.3.3",
-    "vitest": "^3.0.5"
-  }
-}
