--- packages/tests/vitest.config.ts
@@ -1,6 +1,0 @@
-import { defineProject } from "vitest/config";
-
-export default defineProject({
-  test: {
-  }
-});
